pyproject.toml
[tool.poetry]
name = "toggl-cli"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "toggl_cli", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
httpx = "^0.24.1"
python-dotenv = "^1.0.0"
rich = "^13.5.2"
pendulum = "^2.1.2"

[tool.poetry.scripts]
toggl-cli = "toggl_cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
